requirements.txt
grpcio==1.68.0
grpcio-tools==1.68.0
protobuf==5.29.0
grpcio-reflection
torch
numpy
wandb
llama-cpp-python
